+++ src/bes/lims/extender/baseanalysis.py
@@ -19,14 +19,34 @@
 # Some rights reserved, see README and LICENSE.
 
 from archetypes.schemaextender.interfaces import IBrowserLayerAwareExtender
+from archetypes.schemaextender.interfaces import ISchemaExtender
 from archetypes.schemaextender.interfaces import ISchemaModifier
 from bes.lims import messageFactory as _
+from bes.lims.extender.field import ExtBooleanField
 from bes.lims.interfaces import IBESLimsLayer
 from bika.lims import senaiteMessageFactory as _c
 from bika.lims.interfaces import IBaseAnalysis
 from Products.Archetypes import DisplayList
+from Products.Archetypes.Widget import BooleanWidget
 from zope.component import adapter
 from zope.interface import implementer
+
+
+NEW_FIELDS = [
+    ExtBooleanField(
+        "ExcludeFromIntegration",
+        schemata="Analysis",
+        default=False,
+        widget=BooleanWidget(
+            label=_("Exclude from integration"),
+            description=_(
+                "If enabled, the result of this analysis will not be sent to "
+                "integrated systems (e.g. Tamanu). The integrated system will "
+                "receive 'Refer to PDF report' in place of the actual result."
+            ),
+        )
+    ),
+]
 
 
 UPDATED_FIELDS = [
@@ -68,6 +88,18 @@ def update_field(schema, field_id, field_info):
 
 
 @adapter(IBaseAnalysis)
+@implementer(ISchemaExtender, IBrowserLayerAwareExtender)
+class BaseAnalysisSchemaExtender(object):
+    layer = IBESLimsLayer
+
+    def __init__(self, context):
+        self.context = context
+
+    def getFields(self):  # noqa CamelCase
+        return NEW_FIELDS
+
+
+@adapter(IBaseAnalysis)
 @implementer(ISchemaModifier, IBrowserLayerAwareExtender)
 class BaseAnalysisSchemaModifier(object):
     layer = IBESLimsLayer
@@ -79,3 +111,17 @@ class BaseAnalysisSchemaModifier(object):
         for field_name, field_info in UPDATED_FIELDS:
             update_field(schema, field_name, field_info)
         return schema
+
+
+def getExcludeFromIntegration(self):
+    """Returns whether this analysis result should be excluded from
+    integrated systems (e.g. Tamanu)
+    """
+    return self.getField("ExcludeFromIntegration").get(self)
+
+
+def setExcludeFromIntegration(self, value):
+    """Sets whether this analysis result should be excluded from
+    integrated systems (e.g. Tamanu)
+    """
+    self.getField("ExcludeFromIntegration").set(self, value)