+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 1.0.0
 -----
 
+- #158 Add ExcludeFromIntegration field and update diagnostic report handling
 - #125 Fix Tamanu sync freezing due to missing HTTP timeouts
 - #155 Disable notification for sample status transition to 'to_be_verified'
 - #144 Adding reference results to Tamanu FHIR Observations